# schemas/openapi.json
{
  "components": {
    "schemas": {
...
      "AccessJournalEntry": {
        "description": "One access-journal entry (SPEC §2.A, §9.A).",
        "properties": {
          "action": {
            "description": "Stable action name (`pair.window_opened`, `device.paired`,\n`device.revoked`, `auth.rejected`…).",
            "type": "string"
          },
          "at": {
            "description": "When the action happened.",
            "format": "date-time",
            "type": "string"
          },
          "detail": {
            "description": "Non-P0 context (route, device kind…). Never user content.",
            "type": [
              "string",
              "null"
            ]
          },
          "device_id": {
            "description": "Acting device id, when the action was authenticated.",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "at",
          "action"
        ],
        "type": "object"
      },
      "AccessJournalResponse": {
        "description": "`GET /system/journal` — the local access journal, shown in the\ncaregiver space (SPEC §2.A, §7.C). **Metadata only, never P0**\n(SPEC §9.A): entries describe access actions, never the content of a\nconversation, draft, or memory.",
        "properties": {
          "entries": {
            "description": "Recent entries, newest first.",
            "items": {
              "$ref": "#/components/schemas/AccessJournalEntry"
            },
            "type": "array"
          }
        },
        "required": [
          "entries"
        ],
        "type": "object"
      },
      "CapabilitiesResponse": {
        "description": "`GET /system/capabilities` — what this installation can do\n(hardware tier §3, available features).",
        "properties": {
          "api_version": {
            "description": "API version served by this hub (`/api/v1`).",
            "format": "uint32",
            "minimum": 0,
            "type": "integer"
          },
          "features": {
            "description": "Available feature flags (`asr`, `voice_cloning`, `embeddings`…).\nFree-form strings: features appear with releases without a contract\nchange; clients test membership, never enumerate.",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "tier": {
            "$ref": "#/components/schemas/HardwareTier",
            "description": "Hardware tier from installation profiling (SPEC §3)."
          }
        },
        "required": [
          "api_version",
          "tier",
          "features"
        ],
        "type": "object"
      },
...
      "EmergencyRequest": {
        "description": "`POST /system/emergency` — raise or clear the emergency alert (D-7.4).\n\nRequires `control` scope (the user triggers it from the composer after a\ndouble confirmation, SPEC §7.A). The hub broadcasts\n[`SystemEvent::Emergency`] on the `system` topic and rings locally; the\nresulting state reaches the caller through that broadcast, so the call\nitself returns `204 No Content`.",
        "properties": {
          "active": {
            "description": "Desired state: `true` raises the alert, `false` clears it.",
            "type": "boolean"
          }
        },
        "required": [
          "active"
        ],
        "type": "object"
      },
      "ErrorCode": {
        "description": "Stable error codes (wire: `snake_case`).\n\nMarked `non_exhaustive` and tolerant on input ([`ErrorCode::Unknown`]):\nthe catalogue grows; clients fall back on `status` for codes they do not\nknow.",
        "oneOf": [
          {
            "const": "pairing_window_closed",
            "description": "Pairing attempted outside an explicitly opened window (SPEC §2.A).",
            "type": "string"
          },
          {
            "const": "pairing_code_invalid",
            "description": "Wrong or already-consumed pairing code.",
            "type": "string"
          },
          {
            "const": "rate_limited",
            "description": "Too many attempts; retry later (always carried by a 429).",
            "type": "string"
          },
          {
            "const": "token_missing",
            "description": "No `X-Fluence-Token` header on a protected route.",
            "type": "string"
          },
          {
            "const": "token_invalid",
            "description": "Unknown or revoked device token.",
            "type": "string"
          },
          {
            "const": "scope_insufficient",
            "description": "Token scope does not allow this route (SPEC §2.A scope table).",
            "type": "string"
          },
          {
            "const": "session_not_found",
            "description": "Unknown session id.",
            "type": "string"
          },
          {
            "const": "profile_not_found",
            "description": "Unknown profile id.",
            "type": "string"
          },
          {
            "const": "memory_item_not_found",
            "description": "Unknown memory item id.",
            "type": "string"
          },
          {
            "const": "voice_not_found",
            "description": "Unknown voice id.",
            "type": "string"
          },
          {
            "const": "validation_failed",
            "description": "Request shape was valid JSON but violated an invariant.",
            "type": "string"
          },
          {
            "const": "asr_consent_required",
            "description": "ASR control requires a valid consent token (SPEC §5.A).",
            "type": "string"
          },
          {
            "const": "asr_consent_expired",
            "description": "The consent token expired (TTL elapsed).",
            "type": "string"
          },
          {
            "const": "worker_unavailable",
            "description": "The serving worker is down and no fallback applies. Routes with a\nfallback (n-gram, OS voice) degrade instead of erroring (SPEC §2.C).",
            "type": "string"
          },
          {
            "const": "internal",
            "description": "Unexpected internal error (carries no internals in `detail`).",
            "type": "string"
          },
          {
            "const": "unknown",
            "description": "Code added by a newer hub; clients fall back on `status`.",
            "type": "string"
          }
        ]
      },
...
      "HardwareTier": {
        "description": "Hardware tier (SPEC §3 fleet table).",
        "oneOf": [
          {
            "const": "reduced",
            "description": "8 GB RAM, no GPU (FLU-REF-1) — latency budgets are defined here.",
            "type": "string"
          },
          {
            "const": "nominal",
            "description": "16 GB RAM.",
            "type": "string"
          },
          {
            "const": "gpu_hub",
            "description": "≥ 12 GB VRAM GPU hub.",
            "type": "string"
          }
        ]
      },
...
      "Problem": {
        "description": "RFC 9457 problem document, extended with [`ErrorCode`].\n\nThe `type` URI is derived from the code (`urn:fluence:problem:<code>`):\na stable URN that does not depend on owning a domain yet; it may become\na resolvable URL later without changing the codes.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "Stable machine-readable code (the catalogue clients switch on)."
          },
          "detail": {
            "description": "Occurrence-specific explanation. **Never P0 content.**",
            "type": [
              "string",
              "null"
            ]
          },
          "instance": {
            "description": "URI identifying this occurrence (e.g. the request path).",
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "description": "HTTP status code.",
            "format": "uint16",
            "maximum": 65535,
            "minimum": 0,
            "type": "integer"
          },
          "title": {
            "description": "Short human-readable summary, stable per code (English).",
            "type": "string"
          },
          "type": {
            "description": "Stable problem-type URI (`urn:fluence:problem:<code>`).",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "fluenceToken": {
        "description": "Per-device revocable token with a scope (SPEC §2.A)",
        "in": "header",
        "name": "X-Fluence-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Fluence",
      "url": "https://github.com/Pluenet-Killian/Fluence"
    },
    "description": "API of the Fluence hub — a local-first communication platform for people with severe motor disabilities. One session = one conversation with a warm KV-cache; suggestions stream over SSE with per-slot cancellation; input flows through the FluenceInput protocol (WebSocket topics). Authentication: per-device revocable tokens with scopes (display/control/care/system). Operations are marked `x-fluence-stability: stable` (A1 core, frozen) or `experimental` (P2 domains, may change). Errors follow RFC 9457 problem+json with a stable machine-readable code catalogue.",
    "license": {
      "identifier": "Apache-2.0",
      "name": "Apache-2.0"
    },
    "summary": "Local communication hub: input, acceleration, voice (SPEC §5.A)",
    "title": "Fluence hub API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/system/capabilities": {
      "get": {
        "operationId": "get_system_capabilities",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilitiesResponse"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Error (RFC 9457)"
          }
        },
        "security": [
          {
            "fluenceToken": []
          }
        ],
        "summary": "Installation tier and available features",
        "tags": [
          "system"
        ],
        "x-fluence-scopes": [
          "display",
          "control",
          "care"
        ],
        "x-fluence-stability": "stable"
      }
    },
    "/api/v1/system/emergency": {
      "post": {
        "operationId": "post_system_emergency",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmergencyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No content"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Error (RFC 9457)"
          }
        },
        "security": [
          {
            "fluenceToken": []
          }
        ],
        "summary": "Raise or clear the emergency alert (double-confirmed; D-7.4)",
        "tags": [
          "system"
        ],
        "x-fluence-scopes": [
          "control"
        ],
        "x-fluence-stability": "stable"
      }
    },
...
    "/api/v1/system/journal": {
      "get": {
        "operationId": "get_system_journal",
        "parameters": [
          {
            "description": "Maximum entries to return (newest first; default 100)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessJournalResponse"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Error (RFC 9457)"
          }
        },
        "security": [
          {
            "fluenceToken": []
          }
        ],
        "summary": "Local access journal (caregiver space; metadata only)",
        "tags": [
          "system"
        ],
        "x-fluence-scopes": [
          "care"
        ],
        "x-fluence-stability": "stable"
      }
    },
...
  },
  "servers": [
    {
      "description": "Embedded mode (loopback, no TLS)",
      "url": "http://127.0.0.1:7411"
    },
    {
      "description": "Home mode (LAN, local CA)",
      "url": "https://{host}:7411",
      "variables": {
        "host": {
          "default": "fluence.local"
        }
      }
    }
  ],
...
}